Correct usec calculation for tsresol other than 6.
parse_usec returns the self.tsresol least significant integers of the passed time t. The result is the interpreted as microseconds. This is only true, if self.tsresol equals 6. If, for example, tsresol is 9, then an 8-digit number is returned that represents a timestamp in nanoseconds, but is then interpreted as microseconds. This happens for every tsresol other than 6.

This pull request normalizes the return value of parse_usec by pow(10, self.tsresol-6), ensuring that the portion of the timestamp smaller than a second is correctly returned in microsecond format. Thus, a nanosecond (tsresol=9) timescale will return 3 decimal spaces, while a millisecond timescale (tsresol=3) will append three zeros.